Vilkårsvurdering: trekk ut funksjonalitet for å generere endringstidslinje pr person og vilkår
